+++ praw/models/reddit/subreddit.py
@@ -119,7 +119,7 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
 
     @property
     def banned(self):
-        """An instance of :class:`.SubredditRelationship`.
+        """Provide an instance of :class:`.SubredditRelationship`.
 
         For example to ban a user try:
 
@@ -141,21 +141,21 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
 
     @property
     def contributor(self):
-        """An instance of :class:`.ContributorRelationship`."""
+        """Provide an instance of :class:`.ContributorRelationship`."""
         if self._contributor is None:
             self._contributor = ContributorRelationship(self, 'contributor')
         return self._contributor
 
     @property
     def filters(self):
-        """An instance of :class:`.SubredditFilters`."""
+        """Provide an instance of :class:`.SubredditFilters`."""
         if self._filters is None:
             self._filters = SubredditFilters(self)
         return self._filters
 
     @property
     def flair(self):
-        """An instance of :class:`.SubredditFlair`.
+        """Provide an instance of :class:`.SubredditFlair`.
 
         Provides the interface for interacting with a subreddit's flair. For
         example to list all the flair for a subreddit which you have the
@@ -173,14 +173,14 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
 
     @property
     def mod(self):
-        """An instance of :class:`.SubredditModeration`."""
+        """Provide an instance of :class:`.SubredditModeration`."""
         if self._mod is None:
             self._mod = SubredditModeration(self)
         return self._mod
 
     @property
     def moderator(self):
-        """An instance of :class:`.ModeratorRelationship`.
+        """Provide an instance of :class:`.ModeratorRelationship`.
 
         For example to add a moderator try:
 
@@ -202,21 +202,21 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
 
     @property
     def modmail(self):
-        """An instance of :class:`.Modmail`."""
+        """Provide an instance of :class:`.Modmail`."""
         if self._modmail is None:
             self._modmail = Modmail(self)
         return self._modmail
 
     @property
     def muted(self):
-        """An instance of :class:`.SubredditRelationship`."""
+        """Provide an instance of :class:`.SubredditRelationship`."""
         if self._muted is None:
             self._muted = SubredditRelationship(self, 'muted')
         return self._muted
 
     @property
     def quaran(self):
-        """An instance of :class:`.SubredditQuarantine`.
+        """Provide an instance of :class:`.SubredditQuarantine`.
 
         This property is named ``quaran`` because ``quarantine`` is a
         Subreddit attribute returned by Reddit to indicate whether or not a
@@ -229,7 +229,7 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
 
     @property
     def stream(self):
-        """An instance of :class:`.SubredditStream`.
+        """Provide an instance of :class:`.SubredditStream`.
 
         Streams can be used to indefinitely retrieve new comments made to a
         subreddit, like:
@@ -255,14 +255,14 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
 
     @property
     def stylesheet(self):
-        """An instance of :class:`.SubredditStylesheet`."""
+        """Provide an instance of :class:`.SubredditStylesheet`."""
         if self._stylesheet is None:
             self._stylesheet = SubredditStylesheet(self)
         return self._stylesheet
 
     @property
     def wiki(self):
-        """An instance of :class:`.SubredditWiki`.
+        """Provide an instance of :class:`.SubredditWiki`.
 
         This attribute can be used to discover all wikipages for a subreddit:
 
@@ -329,7 +329,7 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
         """
         return self._reddit.get(API_PATH['rules'].format(subreddit=self))
 
-    def search(self, query, sort='relevance', syntax='cloudsearch',
+    def search(self, query, sort='relevance', syntax='lucene',
                time_filter='all', **generator_kwargs):
         """Return a ListingGenerator for items that match ``query``.
 
@@ -337,13 +337,9 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
         :param sort: Can be one of: relevance, hot, top, new,
             comments. (default: relevance).
         :param syntax: Can be one of: cloudsearch, lucene, plain
-            (default: cloudsearch -- will be lucene in PRAW 5).
+            (default: lucene).
         :param time_filter: Can be one of: all, day, hour, month, week, year
             (default: all).
-
-        .. warning:: (Deprecation) The default search syntax is changing to
-           ``lucene`` in PRAW 5. For forward compatibility please explicitly
-           provide the search syntax.
 
         For more information on building a search query see:
             https://www.reddit.com/wiki/search
@@ -427,7 +423,7 @@ class Subreddit(RedditBase, MessageableMixin, SubredditListingMixin):
             current_ids = set()
             found_new_submission = False
             for submission in self.search(query, limit=None, params=params,
-                                          sort='new'):
+                                          sort='new', syntax='cloudsearch'):
                 current_ids.add(submission.id)
                 end = min(end, int(submission.created))
                 if submission.id not in last_ids:
@@ -619,20 +615,6 @@ class SubredditFlair(object):
         self.subreddit = subreddit
         self.templates = SubredditFlairTemplates(subreddit)
 
-    def __iter__(self):
-        """Iterate through the Redditors and their associated flair.
-
-        .. warning:: (Deprecated) This method will be removed in PRAW 5. Prefer
-                     iterating over ``subreddit.flair`` instead of
-                     ``subreddit.flair``.
-
-        """
-        url = API_PATH['flairlist'].format(subreddit=self.subreddit)
-        params = {'unique': self.subreddit._reddit._next_unique}
-        for item in ListingGenerator(self.subreddit._reddit, url, None,
-                                     params=params):
-            yield item
-
     def configure(self, position='right', self_assign=False,
                   link_position='left', link_self_assign=False,
                   **settings):
@@ -671,8 +653,7 @@ class SubredditFlair(object):
 
         """
         url = API_PATH['deleteflair'].format(subreddit=self.subreddit)
-        # PRAW5 REMOVE (return statement)
-        return self.subreddit._reddit.post(url, data={'name': str(redditor)})
+        self.subreddit._reddit.post(url, data={'name': str(redditor)})
 
     def delete_all(self):
         """Delete all Redditor flair in the Subreddit.
@@ -683,7 +664,7 @@ class SubredditFlair(object):
         """
         return self.update(x['user'] for x in self())
 
-    def set(self, redditor=None, text='', css_class='', thing=None):
+    def set(self, redditor=None, text='', css_class=''):
         """Set flair for a Redditor.
 
         :param redditor: (Required) A redditor name (e.g., ``'spez'``) or
@@ -696,13 +677,6 @@ class SubredditFlair(object):
         This method can only be used by an authenticated user who is a
         moderator of the associated Subreddit.
 
-        .. warning:: The use of this method to set the flair of a
-                     :class:`.Submission` is deprecated and will be removed in
-                     PRAW 5. Use :meth:`.flair` instead.
-
-        .. warning:: The use of the keyword argument ``thing`` is deprecated
-                     and will be removed in PRAW 5. Use ``redditor`` instead.
-
         Example:
 
         .. code:: python
@@ -710,17 +684,7 @@ class SubredditFlair(object):
            reddit.subreddit('redditdev').flair.set('bboe', 'PRAW author')
 
         """
-        # PRAW5 REMOVE
-        if bool(redditor) == bool(thing):
-            raise TypeError('`redditor` must be provided.')
-        if redditor is None:
-            redditor = thing
-
-        data = {'css_class': css_class, 'text': text}
-        if redditor.__class__.__name__ == 'Submission':  # PRAW5 REMOVE
-            data['link'] = redditor.fullname
-        else:
-            data['name'] = str(redditor)
+        data = {'css_class': css_class, 'name': str(redditor), 'text': text}
         url = API_PATH['flair'].format(subreddit=self.subreddit)
         self.subreddit._reddit.post(url, data=data)
 
@@ -743,8 +707,8 @@ class SubredditFlair(object):
         :returns: List of dictionaries indicating the success or failure of
             each update.
 
-        For example to clear the flair text, and set the `praw` flair css class
-        on a few users try:
+        For example to clear the flair text, and set the ``praw`` flair css
+        class on a few users try:
 
         .. code:: python
 
@@ -866,33 +830,11 @@ class SubredditModeration(object):
         url = API_PATH['accept_mod_invite'].format(subreddit=self.subreddit)
         self.subreddit._reddit.post(url)
 
-    @staticmethod
-    def approve(thing):
-        """DEPRECATED.
-
-        .. warning:: (Deprecated) This method will be removed in PRAW 5. Prefer
-                     calling ``comment.mod.approve()``,
-                     ``submission.mod.approve()``.
-
-        """
-        thing.mod.approve()
-
-    @staticmethod
-    def distinguish(thing, how='yes', sticky=False):
-        """DEPRECATED.
-
-        .. warning:: (Deprecated) This method will be removed in PRAW 5. Prefer
-                     calling ``comment.mod.distinguish()``,
-                     ``submission.mod.distinguish()``.
-
-        """
-        thing.mod.distinguish(how=how, sticky=sticky)
-
     def edited(self, only=None, **generator_kwargs):
         """Return a ListingGenerator for edited comments and submissions.
 
-        :param only: If specified, one of `comments`, or 'submissions' to yield
-            only results of that type.
+        :param only: If specified, one of ``'comments'``, or ``'submissions'``
+            to yield only results of that type.
 
         Additional keyword arguments are passed in the initialization of
         :class:`.ListingGenerator`.
@@ -909,17 +851,6 @@ class SubredditModeration(object):
         return ListingGenerator(
             self.subreddit._reddit, API_PATH['about_edited'].format(
                 subreddit=self.subreddit), **generator_kwargs)
-
-    @staticmethod
-    def ignore_reports(thing):
-        """DEPRECATED.
-
-        .. warning:: (Deprecated) This method will be removed in PRAW 5. Prefer
-                     calling ``comment.mod.ignore_reports()``,
-                     ``submission.mod.ignore_reports()``.
-
-        """
-        thing.mod.ignore_reports()
 
     def inbox(self, **generator_kwargs):
         """Return a ListingGenerator for moderator messages.
@@ -969,8 +900,8 @@ class SubredditModeration(object):
     def modqueue(self, only=None, **generator_kwargs):
         """Return a ListingGenerator for comments/submissions in the modqueue.
 
-        :param only: If specified, one of `comments`, or 'submissions' to yield
-            only results of that type.
+        :param only: If specified, one of ``'comments'``, or ``'submissions'``
+            to yield only results of that type.
 
         Additional keyword arguments are passed in the initialization of
         :class:`.ListingGenerator`.
@@ -988,22 +919,11 @@ class SubredditModeration(object):
             self.subreddit._reddit, API_PATH['about_modqueue'].format(
                 subreddit=self.subreddit), **generator_kwargs)
 
-    @staticmethod
-    def remove(thing, spam=False):
-        """DEPRECATED.
-
-        .. warning:: (Deprecated) This method will be removed in PRAW 5. Prefer
-                     calling ``comment.mod.remove()``,
-                     ``submission.mod.remove()``.
-
-        """
-        thing.mod.remove(spam=spam)
-
     def reports(self, only=None, **generator_kwargs):
         """Return a ListingGenerator for reported comments and submissions.
 
-        :param only: If specified, one of `comments`, or 'submissions' to yield
-            only results of that type.
+        :param only: If specified, one of ``'comments'``, or ``'submissions'``
+            to yield only results of that type.
 
         Additional keyword arguments are passed in the initialization of
         :class:`.ListingGenerator`.
@@ -1030,8 +950,8 @@ class SubredditModeration(object):
     def spam(self, only=None, **generator_kwargs):
         """Return a ListingGenerator for spam comments and submissions.
 
-        :param only: If specified, one of `comments`, or 'submissions' to yield
-            only results of that type.
+        :param only: If specified, one of ``'comments'``, or ``'submissions'``
+            to yield only results of that type.
 
         Additional keyword arguments are passed in the initialization of
         :class:`.ListingGenerator`.
@@ -1048,28 +968,6 @@ class SubredditModeration(object):
         return ListingGenerator(
             self.subreddit._reddit, API_PATH['about_spam'].format(
                 subreddit=self.subreddit), **generator_kwargs)
-
-    @staticmethod
-    def undistinguish(thing):
-        """DEPRECATED.
-
-        .. warning:: (Deprecated) This method will be removed in PRAW 5. Prefer
-                     calling ``comment.mod.undistinguish()``,
-                     ``submission.mod.undistinguish()``.
-
-        """
-        thing.mod.undistinguish()
-
-    @staticmethod
-    def unignore_reports(thing):
-        """DEPRECATED.
-
-        .. warning:: (Deprecated) This method will be removed in PRAW 5. Prefer
-                     calling ``comment.mod.unignore_reports()``,
-                     ``submission.mod.unignore_reports()``.
-
-        """
-        thing.mod.unignore_reports()
 
     def unmoderated(self, **generator_kwargs):
         """Return a ListingGenerator for unmoderated submissions.
@@ -1128,8 +1026,8 @@ class SubredditModeration(object):
         :param header_hover_text: The text seen when hovering over the snoo.
         :param hide_ads: Don't show ads within this subreddit. Only applies to
             gold-user only subreddits.
-        :param key_color: A 6-digit rgb hex color (e.g. `#AABBCC`), used as a
-            thematic color for your subreddit on mobile.
+        :param key_color: A 6-digit rgb hex color (e.g. ``'#AABBCC'``), used as
+            a thematic color for your subreddit on mobile.
         :param lang: A valid IETF language tag (underscore separated).
         :param link_type: The types of submissions users can make.
             One of ``any``, ``link``, ``self``.
@@ -1197,7 +1095,7 @@ class SubredditModeration(object):
 
 
 class SubredditQuarantine(object):
-    """Provides submission and comment streams."""
+    """Provides subreddit quarantine related methods."""
 
     def __init__(self, subreddit):
         """Create a SubredditQuarantine instance.
@@ -1293,20 +1191,6 @@ class SubredditRelationship(object):
         """
         self.relationship = relationship
         self.subreddit = subreddit
-
-    def __iter__(self):
-        """Iterate through the Redditors belonging to this relationship.
-
-        .. warning:: (Deprecated) This method will be removed in PRAW 5. Prefer
-                     calling ``subreddit.banned(limit=None)`` instead of
-                     ``subreddit.banned`` and similar for other relationships.
-
-        """
-        url = API_PATH['list_{}'.format(self.relationship)].format(
-            subreddit=self.subreddit)
-        params = {'unique': self.subreddit._reddit._next_unique}
-        for item in self.subreddit._reddit.get(url, params=params):
-            yield item
 
     def add(self, redditor, **other_settings):
         """Add ``redditor`` to this relationship.
@@ -1408,10 +1292,10 @@ class ModeratorRelationship(SubredditRelationship):
 
         :param redditor: A redditor name (e.g., ``'spez'``) or
             :class:`~.Redditor` instance.
-        :param permissions: When provided (not `None`), permissions should be a
-            list of strings specifying which subset of permissions to grant. An
-            empty list `[]` indicates no permissions, and when not provided
-            `None`, indicates full permissions.
+        :param permissions: When provided (not ``None``), permissions should be
+            a list of strings specifying which subset of permissions to
+            grant. An empty list ``[]`` indicates no permissions, and when not
+            provided ``None``, indicates full permissions.
 
         An invite will be sent unless the user making this call is an admin
         user.
@@ -1432,10 +1316,10 @@ class ModeratorRelationship(SubredditRelationship):
 
         :param redditor: A redditor name (e.g., ``'spez'``) or
             :class:`~.Redditor` instance.
-        :param permissions: When provided (not `None`), permissions should be a
-            list of strings specifying which subset of permissions to grant. An
-            empty list `[]` indicates no permissions, and when not provided
-            `None`, indicates full permissions.
+        :param permissions: When provided (not ``None``), permissions should be
+            a list of strings specifying which subset of permissions to
+            grant. An empty list ``[]`` indicates no permissions, and when not
+            provided ``None``, indicates full permissions.
 
         For example, to invite ``'spez'`` with ``'posts'`` and ``'mail'``
             permissions to ``'/r/test/``, try:
@@ -1485,10 +1369,10 @@ class ModeratorRelationship(SubredditRelationship):
 
         :param redditor: A redditor name (e.g., ``'spez'``) or
             :class:`~.Redditor` instance.
-        :param permissions: When provided (not `None`), permissions should be a
-            list of strings specifying which subset of permissions to grant. An
-            empty list `[]` indicates no permissions, and when not provided,
-            `None`, indicates full permissions.
+        :param permissions: When provided (not ``None``), permissions should be
+            a list of strings specifying which subset of permissions to
+            grant. An empty list ``[]`` indicates no permissions, and when not
+            provided, ``None``, indicates full permissions.
 
         For example, to add all permissions to the moderator, try:
 
@@ -1513,10 +1397,10 @@ class ModeratorRelationship(SubredditRelationship):
 
         :param redditor: A redditor name (e.g., ``'spez'``) or
             :class:`~.Redditor` instance.
-        :param permissions: When provided (not `None`), permissions should be a
-            list of strings specifying which subset of permissions to grant. An
-            empty list `[]` indicates no permissions, and when not provided,
-            `None`, indicates full permissions.
+        :param permissions: When provided (not ``None``), permissions should be
+            a list of strings specifying which subset of permissions to
+            grant. An empty list ``[]`` indicates no permissions, and when not
+            provided, ``None``, indicates full permissions.
 
         For example, to grant the flair and mail permissions to the moderator
         invite, try:
@@ -1569,8 +1453,8 @@ class Modmail(object):
         this method. Instead, use :meth:`~.Modmail.subreddits` to get a list of
         subreddits using the new modmail.
 
-        :param other_subreddits: A list of `.Subreddit` instances for which to
-            mark conversations (default: None).
+        :param other_subreddits: A list of :class:`.Subreddit` instances for
+            which to mark conversations (default: None).
         :param state: Can be one of: all, archived, highlighted, inprogress,
             mod, new, notifications, (default: all).
         :returns: A list of :class:`.ModmailConversation` instances that were
@@ -1601,8 +1485,8 @@ class Modmail(object):
         :param limit: The maximum number of conversations to fetch. If None,
             the server-side default is 25 at the time of writing
             (default: None).
-        :param other_subreddits: A list of `.Subreddit` instances for which to
-            fetch conversations (default: None).
+        :param other_subreddits: A list of :class:`.Subreddit` instances for
+            which to fetch conversations (default: None).
         :param sort: Can be one of: mod, recent, unread, user
             (default: recent).
         :param state: Can be one of: all, archived, highlighted, inprogress,
@@ -1641,7 +1525,7 @@ class Modmail(object):
             :class:`.Redditor`.
         :param author_hidden: When True, author is hidden from non-moderators
             (default: False).
-        :returns: A `.ModmailConversation` object for the newly created
+        :returns: A :class:`.ModmailConversation` object for the newly created
             conversation.
 
         .. code:: python
@@ -1714,7 +1598,7 @@ class SubredditStream(object):
         Comments are yielded oldest first. Up to 100 historical comments will
         initially be returned.
 
-        Additional keyword arguments are passed to the stream generator.
+        Keyword arguments are passed to :meth:`.stream_generator`.
 
         For example, to retrieve all new comments made to the ``iama``
         subreddit, try:
@@ -1779,11 +1663,13 @@ class SubredditStream(object):
         """
         return stream_generator(self.subreddit.comments, **stream_options)
 
-    def submissions(self):
+    def submissions(self, **stream_options):
         """Yield new submissions as they become available.
 
         Submissions are yielded oldest first. Up to 100 historical submissions
         will initially be returned.
+
+        Keyword arguments are passed to :meth:`.stream_generator`.
 
         For example to retrieve all new submissions made to all of Reddit, try:
 
@@ -1793,7 +1679,7 @@ class SubredditStream(object):
                print(submission)
 
         """
-        return stream_generator(self.subreddit.new)
+        return stream_generator(self.subreddit.new, **stream_options)
 
 
 class SubredditStylesheet(object):